requirements.txt
langchain
langchain_community
langchain_huggingface
faiss-cpu
sentence-transformers
PyMuPDF
tqdm
streamlit
python-dotenv